Use hugoShortcode to highlight `{{ highlight }}` shortcode
...like the other shortcodes.

Here's how it works.

We give the patterns for the start and end of the regions, but don't
highlight them as a single thing with `matchgroup`. Instead, we give an
pattern offset (see `:h syn-pattern-offset`) of ms=s-1 so we can modify
what is highlighted. What I want is to not highlight this part with a
group, but rather highlight it with the other groups already defined.
This offset achieves that by excluding the whole text from being
highlighted with matchgroup.

Then, we say this region contains hugoShortcode, which is the group we
want to use for highlighting.